chore: remove type-checking
